VACMS-14025 Move profile logic from va_gov_backend to va_gov_profile.
