+++ docroot/modules/custom/va_gov_backend/src/EventSubscriber/EntityEventSubscriber.php
@@ -8,7 +8,6 @@ use Drupal\core_event_dispatcher\Event\Entity\EntityTypeBuildEvent;
 use Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent;
 use Drupal\Core\Config\Entity\ConfigEntityType;
 use Drupal\Core\Entity\EntityTypeManager;
-use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\field_event_dispatcher\Event\Field\WidgetSingleElementFormAlterEvent;
 use Drupal\field_event_dispatcher\FieldHookEvents;
@@ -61,8 +60,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
   public static function getSubscribedEvents(): array {
     return [
       'hook_event_dispatcher.form_node_centralized_content_edit_form.alter' => 'alterCentralizedContentNodeForm',
-      'hook_event_dispatcher.form_node_person_profile_edit_form.alter' => 'alterStaffProfileNodeForm',
-      'hook_event_dispatcher.form_node_person_profile_form.alter' => 'alterStaffProfileNodeForm',
       EntityHookEvents::ENTITY_PRE_SAVE => 'entityPresave',
       EntityHookEvents::ENTITY_TYPE_BUILD => 'entityTypeBuild',
       FieldHookEvents::WIDGET_SINGLE_ELEMENT_FORM_ALTER => 'formWidgetAlter',
@@ -123,66 +120,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
     $title = preg_replace('/^\s+/', '', $title);
     $title = preg_replace('/\s+$/', '', $title);
     $node->setTitle($title);
-  }
-
-  /**
-   * Form alterations for staff profile content type.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function alterstaffProfileNodeForm(FormIdAlterEvent $event): void {
-    $this->addStateManagementToBioFields($event);
-  }
-
-  /**
-   * Add states management to bio fields to determine visibility based on bool.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function addStateManagementToBioFields(FormIdAlterEvent $event) {
-    $form = &$event->getForm();
-    $form['#submit'][] = 'Drupal\va_gov_backend\EventSubscriber\EntityEventSubscriber::addStateManagementToBioFieldsSubmitHandler';
-    $form['#attached']['library'][] = 'va_gov_backend/set_body_to_required';
-    $selector = ':input[name="field_complete_biography_create[value]"]';
-    $form['field_intro_text']['widget'][0]['value']['#states'] = [
-      'required' => [
-          [$selector => ['checked' => TRUE]],
-      ],
-      'visible' => [
-            [$selector => ['checked' => TRUE]],
-      ],
-    ];
-    $form['field_body']['#states'] = [
-      'visible' => [
-          [$selector => ['checked' => TRUE]],
-      ],
-    ];
-    $form['field_complete_biography']['#states'] = [
-      'visible' => [
-          [$selector => ['checked' => TRUE]],
-      ],
-    ];
-  }
-
-  /**
-   * Submit handler for DANSE forms.
-   *
-   * @param array $form
-   *   The form.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   */
-  public static function addStateManagementToBioFieldsSubmitHandler(array $form, FormStateInterface $form_state) {
-    $bio_display = !empty($form_state->getUserInput()['field_complete_biography_create']['value']) ? TRUE : FALSE;
-    if (!$bio_display) {
-
-      $form['field_body']['widget']['#required'] = FALSE;
-      $form['field_body']['widget'][0]['#required'] = FALSE;
-      $form['field_body']['widget'][0]['value']['#required'] = FALSE;
-      //'#limit_validation_errors' => [],
-    }
   }
 
   /**
