+++ docroot/modules/custom/va_gov_backend/va_gov_backend.libraries.yml
@@ -66,9 +66,6 @@ service_location_hours:
   version: 1.x
   js:
     js/service_location_hours.js: {}
-set_body_to_required:
-  js:
-    js/set_body_to_required.js: {}
 smart_date_interactions:
   version: 1.x
   js:
