+++ docroot/modules/custom/va_gov_profile/va_gov_profile.libraries.yml
@@ -1,0 +1,3 @@
+set_body_to_required:
+  js:
+    js/set_body_to_required.js: {}
