+++ docroot/modules/custom/va_gov_profile/va_gov_profile.services.yml
@@ -1,0 +1,6 @@
+services:
+  va_gov_backend.entity_event_subscriber:
+    class: Drupal\va_gov_profile\EventSubscriber\EntityEventSubscriber
+    arguments: ['@va_gov_user.user_perms', '@entity_type.manager']
+    tags:
+      - { name: event_subscriber }
